+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to
 ### Fixed
 
 - 🐛(backend) allow inviting external person on item with no direct access
+- 🐛(backend) stop storing numchild in database use annotation instead
 
 ## [v0.14.0] - 2026-02-25